update on nested loops

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	unsigned int a = 2;
-	unsigned int n = 612852475143;
+	unsigned long n = 612852475143;
 
 	while (a != n)
 	{
